🎉 Fix InfoPanel plugin interface implementation
- Updated BasePlugin interface to match InfoPanel.Plugins v1.0
- Added required abstract methods: Load(), Update(), UpdateAsync(), Close()
- Added UpdateInterval property returning TimeSpan
- Removed deprecated Dispose() override (not part of interface)
- Fixed container sensor registration using containers.Add()
- Added reflection-based sensor addition to discover correct API
- Fixed double-to-float conversion issues in sensor values
- Build now succeeds with complete plugin package output

✅ Release build working: InfoPanel.SteamAPI-v1.0.0.zip created"

+++ InfoPanel.SteamAPI/Services/SensorManagementService.cs
@@ -116,7 +116,7 @@ namespace InfoPanel.SteamAPI.Services
             if (data.IsInGame() && !string.IsNullOrEmpty(data.CurrentGameName))
             {
                 currentGameSensor.Value = data.CurrentGameName;
-                currentGamePlaytimeSensor.Value = Math.Round(data.TotalPlaytimeHours, 1);
+                currentGamePlaytimeSensor.Value = (float)Math.Round(data.TotalPlaytimeHours, 1);
             }
             else
             {
@@ -135,13 +135,13 @@ namespace InfoPanel.SteamAPI.Services
             SteamData data)
         {
             // Update total games owned
-            totalGamesSensor.Value = data.TotalGamesOwned;
+            totalGamesSensor.Value = (float)data.TotalGamesOwned;
             
             // Update total playtime
-            totalPlaytimeSensor.Value = Math.Round(data.TotalLibraryPlaytimeHours, 1);
+            totalPlaytimeSensor.Value = (float)Math.Round(data.TotalLibraryPlaytimeHours, 1);
             
             // Update recent playtime
-            recentPlaytimeSensor.Value = Math.Round(data.RecentPlaytimeHours, 1);
+            recentPlaytimeSensor.Value = (float)Math.Round(data.RecentPlaytimeHours, 1);
         }
         
         /// <summary>
